refactor: 출석 정보 클래스명 변경
